Raspberry Pi 4B/Display OLED/requirements.txt
streamlit
opencv-python
paho-mqtt
pandas
networkx
pyzbar
libzbar-dev
picamera2
libgl1-mesa-glx